ci: fix port in CI
