--- tests/__init__.py
@@ -1,17 +1,0 @@
-import time
-
-import os
-
-import pytest
-
-
-@pytest.fixture()
-def docker_compose(request):
-    os.system(
-        f"docker-compose -f {request.param} --project-directory . up  --build -d --remove-orphans"
-    )
-    time.sleep(5)
-    yield
-    os.system(
-        f"docker-compose -f {request.param} --project-directory . down --remove-orphans"
-    )
